Turret transverse rotation added, interesting problems
The turret now follows the camera more slowly, only on the y rotation axis (side to side). Clamping using EulerAngles was causing terrible problems, will need to figure out how to limit the camera rotation with quaternions and stuff. Like, make it so the player can't look in the opposite direction as the turret while it's moving etc.